+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Remove not needed Paneset to correctly calculate pane widths when closing plugin. Fixes UIPCIR-78.
 * Jest/RTL: Cover `useOptions` hook with unit tests. Refs UIPCIR-68.
 * Jest/RTL: Cover `CreateRecordsWrapper` component with unit tests. Refs UIPCIR-64.
+* Jest/RTL: Cover `CreateRecordsForm` component with unit tests. Refs UIPCIR-66.
 
 ## [4.0.0](https://github.com/folio-org/ui-plugin-create-inventory-records/tree/v4.0.0) (2023-10-12)
 [Full Changelog](https://github.com/folio-org/ui-plugin-create-inventory-records/compare/v3.3.0...v4.0.0)